31/7/2023

+++ app/Http/Controllers/Admin/postController.php
@@ -6,8 +6,11 @@ use App\Http\Controllers\Controller;
 use App\Models\Admin\Category;
 use App\Models\Admin\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Session;
+use App\Models\Admin\Group;
+use Illuminate\Support\Facades\Storage;
 
 class postController extends Controller
 {
@@ -21,7 +24,7 @@ class postController extends Controller
 
     public function index()
     {
-            $posts = Post::paginate(6);
+            $posts = Post::where('status', 1)->paginate(6);
             Session::put('posts_url', request()->fullUrl());
             return view('admin.post.index', compact('posts'));
     }
@@ -29,9 +32,13 @@ class postController extends Controller
 
     public function create()
     {
-        $this->authorize('create', Post::class);
+        if (auth()->user()->can('create-post')) {
         $categories = Category::all();
-        return view ('admin.post.create', compact('categories'));
+        $groups = Group::all();
+        return view ('admin.post.create', compact('categories', 'groups'));
+        } else {
+            return abort(403);
+        }
     }
 
 
@@ -45,6 +52,10 @@ class postController extends Controller
                 'avatar' => 'required',
                 'short_description' => 'required',
                 'is_landing' => 'required',
+                'tags' => 'required',
+                'slug' => 'required',
+                'group_id' => 'required',
+                'avatar' => 'required',
 
             ];
             $messages = [
@@ -56,7 +67,14 @@ class postController extends Controller
             $input = $request->all();
             unset($input['_token']);
             unset($input['category_id']);
+            unset($input['avatar']);
+            $file = $request->file('avatar');
+            $file->storeAs('image' , time().'.'.$file->getClientOriginalExtension(),'public');
+            $image = time().'.'.$file->getClientOriginalExtension();
             $post = Post::create($input);
+            $post->update([
+                'avatar' => $image,
+            ]);
             $post->categories()->attach($request->input('category_id'));
             if (Session::get('post_url')) {
                 return redirect(session('post_url'))->with('success', 'Đã thêm post thành công');
@@ -82,16 +100,22 @@ class postController extends Controller
 
     public function edit($id)
     {
-        $post= Post::find($id);
-        $categories = Category::all();
-        $selectedID = $post->categories->pluck('id')->toArray();
-        $this->authorize('update', $post);
-        return view('admin.post.edit', compact('post','categories','selectedID'));
+
+        if (auth()->user()->can('edit-post')) {
+            $post = Post::find($id);
+            $groups = Group::all();
+            $categories = Category::all();
+            $selectedID = $post->categories->pluck('id')->toArray();
+            return view('admin.post.edit', compact('post', 'categories', 'selectedID', 'groups'));
+        } else {
+            return abort(403);
+        }
     }
 
 
     public function update(Request $request, $id)
     {
+
         if ($request->isMethod('POST')) {
             $rules = [
                 'title' => 'required',
@@ -101,45 +125,67 @@ class postController extends Controller
                 'author' => 'required',
                 'short_description' => 'required',
                 'is_landing' => 'required',
+                'group_id' => 'required',
+                'slug' => 'slug'
             ];
             $messages = [
                 'required' => 'Không được để trống trường này',
             ];
             $request->validate($rules, $messages);
         }
-//        try {
+        try {
             $post = Post::find($id);
             $input = $request->all();
             unset($input['_token']);
             unset($input['category_id']);
+            unset($input['avatar']);
             $post->update($input);
-            $post->categories()->sync($request->input('category_id'));
+            $post->save();
+            if($request->hasFile('avatar')) {
+                $destination = 'storage/image/' . $post->avatar;
+                if (File::exists($destination)) {
+                    Storage::delete($post->avatar);
+                }
+                $file = $request->file('avatar');
+                $file->storeAs('public/image', time() . '.' . $file->getClientOriginalExtension());
+                $image = time() . '.' . $file->getClientOriginalExtension();
+
+                $post->update([
+                    'avatar' => $image,
+                ]);
+                $post->categories()->sync($request->input('category_id'));
+            }
             if (Session::get('post_url')) {
                 return redirect(session('post_url'))->with('success', 'Đã sửa post thành công');
             } else {
                 return redirect('admin/post/index')->with('success', 'Đã sửa post thành công');
             }
-//        }catch (\Exception $e) {
-//            return redirect()->back()->with('error', $e->getMessage());
-//        }
+        }catch (\Exception $e) {
+            return redirect()->back()->with('error', $e->getMessage());
+        }
     }
 
 
     public function destroy($id)
     {
-        try {
-            $post = Post::find($id);
-            $this->authorize('delete', $post);
-            $post->delete();
-            if (Session::get('post_url')) {
-                return redirect(session('post_url'))->with('success', 'Đã xóa post thành công');
+            if (auth()->user()->can('delete-post')) {
+                try {
+                    $post = Post::find($id);
+                    $this->authorize('delete', $post);
+                    $post->delete();
+                    if (Session::get('post_url')) {
+                        return redirect(session('post_url'))->with('success', 'Đã xóa post thành công');
+                    } else {
+                        return redirect('admin/post/index')->with('success', 'Đã xóa post thành công');
+                    }
+                } catch (\Exception $e) {
+                    return redirect()->back()->with('error', $e->getMessage());
+                }
             } else {
-                return redirect('admin/post/index')->with('success', 'Đã xóa post thành công');
+                return abort(403);
             }
-        }catch (\Exception $e) {
-            return redirect()->back()->with('error', $e->getMessage());
-        }
     }
+
 
     public function restore() {
         $posts = Post::onlyTrashed()->paginate(6);
